refactor: ResultContents를 HomeContents 와 비슷하게 코드 개선
